Use random() instead of rand() to increase portability
Quoting glibc rand(3):

The  versions  of rand() and srand() in the Linux C Library use the same random
number generator as random(3) and srandom(3), so the lower-order bits should be
as  random  as  the  higher-order bits.   However,  on  older  rand()
implementations, and on current implementations on different systems, the
lower-order bits are much less random than the higher-order bits.  Do not use
this function in applications intended to be portable when good randomness is
needed.  (Use random(3) instead.)